fix_inf_check_m4: fix ibverbs configuration problem
This fix an issue in the detection of ibverbs for situations in which
the locations of header and libraries are rather inhomogeneous (e.g.,
very different source trees). With the changes, when
ibverbs path is not specified, not only the headers but also the
libraries (used by the preprocessor) are fully checked, and in
ultimate instance let to autotools functions if nothing is found by
the first procedure.