queues: fix clean up of queues in proc_term
During GPI-2 the clean-up phase of `gaspi_proc_term` the created
queues are destroyed, the corresponding loops run over the default
number of queues (8) and *not* over the actual number of queues. This
leads to a segmentation fault problem when, e.g., the number of actual
queues is less than eight *and* infiniband is used. Although the same
does not happen for Ethernet, the commit also include changes in TCP
for consistency.