queues: fix clean up of queues in proc_term
During GPI-2 the clean-up phase of `gaspi_proc_term` the created
queues are destroyed, the corresponding loops run first over extra
queues and then over the default number of queues (8). This leads to a
segmentation fault problem when, e.g., the number of actual queues is
less or greater than eight *and* infiniband is used. Now, the extra
queues are still removed at first, but then the minimum between eight
and the actual queues determines the remaining ones to remove.
Although the same does not happen for Ethernet, the commit also
include changes in TCP for consistency.