infiniband: automatic selection of a *working* IB device
This fix issues (e.g. #35) related to the automatic selection (i.e.,
no specified by the user) of an Infiniband device which, although
present, is, e.g., in a DOWN state. To this end, during initialization
also it is checked that the device can be opened and closed, as well
as, the ports are up.