Post a notification after data is read in passive communication.
This commit fixes the race causing a deadlock in issue #40.
In passive communication, a notification to the sender is
sent as soon as the header for data transfer is received
and not after it has actually been received. The sender after
receiving the notification, posts a work completion request
which unblocks the send operation even before the send being
completed leading to a race.