Fix issue with thread-safety in collectives.
When using different groups, the counter of operations posted to the
respective queue needs to account for the case when collective
operations are performed on different groups (e.g. by different
threads). The handling of the counter is now done with atomic
operations. Tests were added for both barrier and allreduce.